+++ Mk/bin/copyright.py
@@ -5,7 +5,7 @@ NOTICE = """
 xnumon - monitor macOS for malicious activity
 https://www.roe.ch/xnumon
 
-Copyright (c) 2017-2018, Daniel Roethlisberger <daniel@roe.ch>.
+Copyright (c) 2017-2019, Daniel Roethlisberger <daniel@roe.ch>.
 All rights reserved.
 
 Licensed under the Open Software License version 3.0.