+++ NEWS.md
@@ -47,6 +47,16 @@ Event schema changes:
 
 ---
 
+### xnumon 0.1.7.2 2019-01-27
+
+-   Avoid calling file operations that would block while the kext is blocking
+    processes during exec(2); this avoids deadlocking the system while
+    long-running low-level disk operation are in progress, such as while Boot
+    Camp Assistant is editing partitions or Disk Utility is running fsck.
+-   Drop support for OS X 10.11 El Capitan.
+
+---
+
 ### xnumon 0.1.7.1 2018-10-08
 
 -   Bugfix release fixing the handling of ptrace(2) for non-`PT_ATTACHEXC`